Adicionando interface nos serviços para desacoprar/separar responsabilidades das classes. - Ajustandos alguns métodos
