+++ src/main/java/com/mvp/produtoapi/controller/PedidoController.java
@@ -6,7 +6,7 @@ import com.mvp.produtoapi.exception.PedidoIllegalArgumentException;
 import com.mvp.produtoapi.exception.PedidoInternalServerErrorException;
 import com.mvp.produtoapi.exception.PedidoNotFoundException;
 import com.mvp.produtoapi.exception.ProdutoNotFoundException;
-import com.mvp.produtoapi.service.PedidoService;
+import com.mvp.produtoapi.service.PedidoServiceImp;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -24,17 +24,17 @@ import java.util.UUID;
 @RequestMapping("/pedidos")
 public class PedidoController {
 
-    private final PedidoService pedidoService;
+    private final PedidoServiceImp pedidoServiceImp;
 
     @Autowired
-    public PedidoController(PedidoService pedidoService) {
-        this.pedidoService = pedidoService;
+    public PedidoController(PedidoServiceImp pedidoServiceImp) {
+        this.pedidoServiceImp = pedidoServiceImp;
     }
 
     @PostMapping("/adicionar")
     @Transactional
     public ResponseEntity<PedidoDTO> cadastrarPedido(@RequestBody PedidoDTO pedidoDTO, UriComponentsBuilder uriBuilder) {
-        Pedido pedido = pedidoService.cadastrarPedido(pedidoDTO);
+        Pedido pedido = pedidoServiceImp.cadastrarPedido(pedidoDTO);
 
         if (pedido == null) {
             throw new PedidoInternalServerErrorException("Erro ao cadastrar pedido");
@@ -46,7 +46,7 @@ public class PedidoController {
 
     @GetMapping
     public ResponseEntity<List<PedidoDTO>> listarTodosPedidos(@RequestParam int pagina, @RequestParam int quantidade, @RequestParam String ordenacao, @RequestParam String direcao) {
-        Page<Pedido> pedidos = pedidoService.listarTodosPedidos(pagina, quantidade, ordenacao, direcao);
+        Page<Pedido> pedidos = pedidoServiceImp.listarTodosPedidos(pagina, quantidade, ordenacao, direcao);
 
         if (pedidos.isEmpty()) {
             throw new ProdutoNotFoundException("Nenhum pedido encontrado");
@@ -58,7 +58,7 @@ public class PedidoController {
 
     @GetMapping("/{id}")
     public ResponseEntity<PedidoDTO> buscarPedidoPorId(@PathVariable UUID id) {
-        Optional<Pedido> pedidos = pedidoService.buscarPedidoPorId(id);
+        Optional<Pedido> pedidos = pedidoServiceImp.buscarPedidoPorId(id);
 
         if (pedidos.isEmpty()) {
             throw new PedidoNotFoundException("Pedido não encontrado");
@@ -70,26 +70,26 @@ public class PedidoController {
     @PatchMapping("{id}")
     @Transactional
     public ResponseEntity<PedidoDTO> atualizarPedido(@PathVariable UUID id , @RequestBody PedidoDTO pedidoDTO) {
-        Optional<Pedido> pedido = pedidoService.buscarPedidoPorId(id);
+        Optional<Pedido> pedido = pedidoServiceImp.buscarPedidoPorId(id);
 
         if (pedido.isEmpty()) {
             throw new PedidoNotFoundException("Pedido não encontrado");
         }
 
-        pedidoService.atualizarPedido(pedidoDTO);
-        return ResponseEntity.ok(pedidoDTO);
+        Pedido pedidoAtualizado = pedidoServiceImp.atualizarPedido(pedidoDTO);
+        return ResponseEntity.ok(PedidoDTO.converterToDto(pedidoAtualizado));
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<Void> removerPedido(@PathVariable UUID id) {
-        Optional<Pedido> Pedido = pedidoService.buscarPedidoPorId(id);
+        Optional<Pedido> Pedido = pedidoServiceImp.buscarPedidoPorId(id);
 
         if (Pedido.isEmpty()) {
             throw new PedidoNotFoundException("Pedido não encontrado");
         }
 
-        pedidoService.removePedido(id);
+        pedidoServiceImp.removerPedido(id);
         return ResponseEntity.noContent().build();
     }
 
@@ -97,7 +97,7 @@ public class PedidoController {
     @Transactional
     public ResponseEntity<PedidoDTO> aplicarDesconto(@PathVariable UUID id, @PathVariable BigDecimal desconto) {
 
-        Optional<Pedido> pedido = pedidoService.buscarPedidoPorId(id);
+        Optional<Pedido> pedido = pedidoServiceImp.buscarPedidoPorId(id);
         if (pedido.isEmpty()) {
             throw new PedidoNotFoundException("Pedido não encontrado");
         }
@@ -108,7 +108,7 @@ public class PedidoController {
             throw new PedidoIllegalArgumentException("Desconto não pode ser maior que 100%");
         }
 
-        PedidoDTO pedidoDTO = pedidoService.aplicarDesconto(pedido.get(), desconto);
+        PedidoDTO pedidoDTO = pedidoServiceImp.aplicarDesconto(pedido.get(), desconto);
 
         return ResponseEntity.ok(pedidoDTO);
     }
