+++ src/main/java/com/mvp/produtoapi/controller/ProdutoController.java
@@ -5,7 +5,7 @@ import com.mvp.produtoapi.entity.Produto;
 import com.mvp.produtoapi.exception.ProdutoBadRequestException;
 import com.mvp.produtoapi.exception.ProdutoInternalServerErrorException;
 import com.mvp.produtoapi.exception.ProdutoNotFoundException;
-import com.mvp.produtoapi.service.ProdutoService;
+import com.mvp.produtoapi.service.ProdutoServiceImp;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -22,27 +22,27 @@ import java.util.UUID;
 @RequestMapping("/produtos")
 public class ProdutoController {
 
-    private final ProdutoService produtoService;
+    private final ProdutoServiceImp produtoServiceImp;
 
     @Autowired
-    public ProdutoController(ProdutoService produtoService) {
-        this.produtoService = produtoService;
+    public ProdutoController(ProdutoServiceImp produtoServiceImp) {
+        this.produtoServiceImp = produtoServiceImp;
     }
 
     @PostMapping("/adicionar")
     @Transactional
     public ResponseEntity<ProdutoDTO> cadastrarProduto(@RequestBody ProdutoDTO produtoDTO, UriComponentsBuilder uriBuilder) {
-        Produto produto = produtoService.cadastrarProduto(produtoDTO);
+        Produto produto = produtoServiceImp.cadastrarProduto(produtoDTO);
         if (produto == null) {
             throw new ProdutoInternalServerErrorException("Erro ao cadastrar produto");
         }
         URI uri = uriBuilder.path("produtos/{id}").buildAndExpand(produto.getId()).toUri();
-        return ResponseEntity.created(uri).body(produtoDTO.converterToDto(produto));
+        return ResponseEntity.created(uri).body(ProdutoDTO.converterToDto(produto));
     }
 
     @GetMapping
     public ResponseEntity<List<ProdutoDTO>> listarTodosProdutos(@RequestParam int pagina, @RequestParam int quantidade, @RequestParam String ordenacao, @RequestParam String direcao) {
-        Page<Produto> produtos = produtoService.listarTodosProdutos(pagina, quantidade, ordenacao, direcao);
+        Page<Produto> produtos = produtoServiceImp.listarTodosProdutos(pagina, quantidade, ordenacao, direcao);
         if (produtos.isEmpty()) {
             throw new ProdutoNotFoundException("Nenhum produto encontrado");
         }
@@ -52,7 +52,7 @@ public class ProdutoController {
 
     @GetMapping("/{id}")
     public ProdutoDTO buscarProdutoPorId(@PathVariable UUID id) {
-        Optional<Produto> produtos = produtoService.buscarProdutoPorId(id);
+        Optional<Produto> produtos = produtoServiceImp.buscarProdutoPorId(id);
 
         if (produtos.isEmpty()) {
             throw new ProdutoNotFoundException(id);
@@ -64,31 +64,31 @@ public class ProdutoController {
     @PatchMapping("{id}")
     @Transactional
     public ResponseEntity<ProdutoDTO> atualizarProduto(@PathVariable UUID id , @RequestBody ProdutoDTO produtoDTO) {
-        Optional<Produto> produto = produtoService.buscarProdutoPorId(id);
+        Optional<Produto> produto = produtoServiceImp.buscarProdutoPorId(id);
 
         if (produto.isEmpty()) {
             throw new ProdutoNotFoundException(id);
         }
 
-        produtoService.atualizarProduto(produtoDTO);
-        return ResponseEntity.ok(produtoDTO);
+        Produto produtoAtualizado = produtoServiceImp.atualizarProduto(produtoDTO);
+        return ResponseEntity.ok(ProdutoDTO.converterToDto(produtoAtualizado));
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<String> removerProduto(@PathVariable UUID id) {
-        Optional<Produto> produto = produtoService.buscarProdutoPorId(id);
+        Optional<Produto> produto = produtoServiceImp.buscarProdutoPorId(id);
 
         if (produto.isEmpty()) {
             throw new ProdutoNotFoundException(id);
         }
 
-        Boolean estaEmPedido = produtoService.verificarSeProdutoEstaEmPedido(id);
+        Boolean estaEmPedido = produtoServiceImp.verificarSeProdutoEstaEmPedido(id);
         if (estaEmPedido) {
             throw new ProdutoBadRequestException(id, "Produto não pode ser removido pois está em um pedido, id: ");
         }
 
-        produtoService.removeProduto(id);
+        produtoServiceImp.removerProduto(id);
         return ResponseEntity.noContent().build();
     }
